pyproject.toml
[project]
name = "schedule-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow[postgres]==2.1.3",
]
